+++ playerdo/backends/moc.py
@@ -5,6 +5,7 @@ from playerdo.utils import process_stdout, process_retval
 # Convoluted expressions fpr Python 2/3 compat
 STATE_STOP = "State: STOP".encode('ascii')
 STATE_PAUSE = "State: PAUSE".encode('ascii')
+STATE_PLAY = "State: PLAY".encode('ascii')
 
 
 class Moc(Player):
@@ -19,6 +20,10 @@ class Moc(Player):
     def is_paused(self):
         info = process_stdout(["mocp", "-i"])
         return STATE_PAUSE in info
+
+    def is_playing(self):
+        info = process_stdout(["mocp", "-i"])
+        return STATE_PLAY in info
 
     def play(self):
         process_retval(["mocp", "--play"])