+++ playerdo/backends/shellfm.py
@@ -15,10 +15,18 @@ class ShellFm(SocketPlayerMixin, Player):
         conf = open(rc_path).read()
         return re.search(r'^\s*unix\s*=\s*([^#\s]+)', conf, re.MULTILINE).groups()[0]
 
+    def _get_status(self):
+        # This only works with shell-fm 0.8 and greater
+        return self.send_socket_command("status", receive=10000).strip()
+
     def is_stopped(self):
-        return not os.path.isfile(os.path.join(os.environ['HOME'],
-                                               ".shell-fm",
-                                               "nowplaying"))
+        return self._get_status() == "STOPPED"
+
+    def is_paused(self):
+        return self._get_status() == "PAUSED"
+
+    def is_playing(self):
+        return self._get_status() == "PLAYING"
 
     def check_dependencies(self):
         retval = []
@@ -35,10 +43,12 @@ class ShellFm(SocketPlayerMixin, Player):
             raise PlayerException("Cannot play shell-fm when in a stopped state.")
 
     def pause(self):
-        self.send_socket_command("pause")
+        if not self.is_paused():
+            self.send_socket_command("pause")
 
     def unpause(self):
-        self.pause()
+        if self.is_paused():
+            self.send_socket_command("pause")
 
     def stop(self):
         self.send_socket_command("stop")